pulled the DataEntryService base class up to its (sole) inheritor
